feat(keycloak): dynamic Keycloak realm and DID provisioning

+++ backend/src/service/registration.service.ts
@@ -1,3 +1,4 @@
+import { Registration } from "../entity/registration.entity";
 import { TrusterIssuer } from "../type/truster-issuer";
 import { keycloakService } from "./keycloak.service";
 import { logger } from "./logger";
@@ -5,22 +6,28 @@ import { tirService } from "./tir.service";
 
 class RegistrationService {
 
-    async register(did: string) {
-        // await keycloakService.createRealm(did);
+    async register(registration: Registration) {
+        if (registration.didGenerated) {
+            await keycloakService.createRealm(registration);
+        }
         const tirIssuer: TrusterIssuer = {
-            did,
+            did: registration.did,
             credentials: []
         }
         try {
             await tirService.registerDid(tirIssuer);
         } catch(error) {
             logger.info('Remove realm because register DID failed');
-            await keycloakService.removeRealm(did);
+            if (registration.didGenerated) {
+                await keycloakService.removeRealm(registration.did);
+            }
             throw error;
         }
     }
-    async unregister(did: string) {
-        // await keycloakService.removeRealm(did);
+    async unregister(did: string, removeRealm: boolean = false) {
+        if (removeRealm) {
+            await keycloakService.removeRealm(did);
+        }
         await tirService.deleteDid(did);
     }
 }
